phase_iii/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart
pydantic>=2.5.0
openai>=1.0.0
python-dotenv
streamlit>=1.28.0
mcp>=1.0.0